requirements.txt
praw==7.4.0
googlesearch_python==1.0.1
requests==2.25.1
discord.py==1.7.3
beautifulsoup4==4.9.3
discord==1.7.3
python-dotenv==0.18.0
wikipedia==1.4.0